feat: allowance list skelleton
