+++ frontend/pages/home/components/AllowanceList.tsx
@@ -1,0 +1,5 @@
+import Table from "@components/Table";
+
+export default function AllowanceList() {
+  return <Table />;
+}
