+++ frontend/pages/home/components/DetailTab.tsx
@@ -1,16 +1,17 @@
 import Button from "@components/buttons/Button";
 import PlusIcon from "@assets/svg/files/plus-icon.svg";
+import { DetailsTabs } from "@/const";
 
 interface Props {
   activeTab: string;
-  setActiveTab: (tab: string) => void;
+  setActiveTab: (tab: DetailsTabs) => void;
 }
 
 export default function DetailTab({ activeTab, setActiveTab }: Props) {
   return (
     <div className="flex justify-between items-center w-full">
-      <div className="columns-2">
-        <button onClick={() => setActiveTab("transactions")}>
+      <div className="columns-2 flex justify-between items-center">
+        <button onClick={() => setActiveTab(DetailsTabs.transactions)}>
           <p
             className={`text-md ${
               activeTab === "transactions" && "border-b-2 border-acceptButtonColor font-semibold text-acceptButtonColor"
@@ -19,7 +20,8 @@ export default function DetailTab({ activeTab, setActiveTab }: Props) {
             Transactions (10)
           </p>
         </button>
-        <button onClick={() => setActiveTab("allowances")}>
+
+        <button onClick={() => setActiveTab(DetailsTabs.allowances)}>
           <p
             className={`text-md ${
               activeTab === "allowances" && "border-b-2 border-acceptButtonColor font-semibold text-acceptButtonColor"
